refactor: isolating error handling
