+++ src/components/SelectActivities/index.js
@@ -3,11 +3,7 @@ import { Typography } from '@material-ui/core';
 import { useEffect } from 'react';
 import { useState } from 'react';
 import useTicket from '../../hooks/api/useTicket';
-
-export const TicketStatus = Object.freeze({
-  PAID: 'PAID',
-  RESERVED: 'RESERVED',
-});
+import { verifyPermission } from './verifyPermission';
 
 export default function SelectActivities() {
   const { ticket, ticketError } = useTicket();
@@ -17,30 +13,8 @@ export default function SelectActivities() {
   });
 
   useEffect(() => {
-    if (ticket) verifyPermission();
-    if (ticketError) {
-      setAutStatus({
-        isAllowed: false,
-        message: 'Você precisa ter selecionar seu ingresso antes de fazer a escolha de atividades',
-      });
-    }
+    if (ticket || ticketError) verifyPermission({ ticket, setAutStatus, ticketError });
   }, [ticket, ticketError]);
-
-  function verifyPermission() {
-    if (ticket.status === TicketStatus.RESERVED) {
-      setAutStatus({
-        isAllowed: false,
-        message: 'Você precisa ter confirmado pagamento antes de fazer a escolha de atividades',
-      });
-    }
-
-    if (ticket.TicketType.isRemote === true) {
-      setAutStatus({
-        isAllowed: false,
-        message: 'Sua modalidade de ingresso não necessita escolher atividade. Você terá acesso a todas as atividades.',
-      });
-    }
-  }
 
   return (
     <>
